Prac10_Solar_Data_Analysis/DB_Extraction/requirements.txt
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
